hw1_lauton04
This assignment took me about an hour to complete, most of which
involved getting used to the Canopy environment and getting a feel for
Python.  The readings, tutorials and lecture included most of what I
needed to know to solve the problems, but a bit of trial and error and
experimentation were needed as well.