Songwriting Project pt. 2
Added classes with pitch info for Chords, Melody, and Bass.

Modified the code a bit (in the Drums class but soon to be in the other
3 parts as well) so that it's easier to change the number of beats.
Specific numbers were replaced with sectionLength and instruments.length
integer variables.

Added an "All" class that opens all 4 parts simultaneously.

Added Clear buttons to uncheck all checkboxes.

Added Save, Load, and Stop buttons, but they do not work yet.

Added a text field that will be used to display the tempo, but it does
not work yet.